backend.oai: use API key to access Proaixy
