+++ backend/oai.py
@@ -23,23 +23,25 @@ from __future__ import unicode_literals
 from datetime import datetime
 import itertools
 import re
+import json
 
 from django.db import transaction
 
-# metadat readers
 from backend import crossref
 from backend.crossref import CrossRefAPI
 from backend.extractors import *
 from backend.name_cache import name_lookup_cache
 from backend.papersource import *
-from backend.proxy import *
 from backend.pubtype_translations import OAI_PUBTYPE_TRANSLATIONS
+from django.conf import settings
+from oaipmh import common
 from oaipmh.client import Client
 from oaipmh.error import BadArgumentError
 from oaipmh.error import DatestampError
 from oaipmh.error import NoRecordsMatchError
 from oaipmh.metadata import base_dc_reader
 from oaipmh.metadata import MetadataRegistry
+from oaipmh.metadata import MetadataReader
 from oaipmh.metadata import oai_dc_reader
 from papers.baremodels import BareName
 from papers.baremodels import BareOaiRecord
@@ -54,6 +56,8 @@ from papers.name import parse_comma_name
 from papers.utils import sanitize_html
 from papers.utils import tolerant_datestamp_to_datetime
 
+# Set exposed by proaixy to indicate the metadata source
+PROXY_SOURCE_PREFIX = "proaixy:source:"
 
 class OaiTranslator(object):
     """
@@ -78,6 +82,19 @@ class OaiTranslator(object):
         """
         raise NotImplemented
 
+class CiteprocReader(MetadataReader):
+
+    def __init__(self):
+        super(CiteprocReader, self).__init__({}, {})
+
+    def __call__(self, element):
+        # extract the Json
+        jsontxt = element.text
+        payload = json.loads(jsontxt)
+
+        return common.Metadata(element, payload)
+
+citeproc_reader = CiteprocReader()
 
 class CiteprocTranslator(object):
     """
@@ -291,6 +308,8 @@ class OaiPaperSource(PaperSource):  # TODO: this should not inherit from PaperSo
         self.registry.registerReader('citeproc', citeproc_reader)
         self.client = Client(endpoint, self.registry)
         self.client._day_granularity = day_granularity
+        self.client.extra_parameters = {
+            'key':settings.PROAIXY_API_KEY }
         self.translators = {}
 
     # Translator management
